Create an encoding class to avoid dual-coding.
Rather than encoding information about a pseudocode function twice
(once in a file like `arithmetic.py` and once at the packages-level
`__init__.py` file), use a decorator class as well as some clever
introspection to encode this information once: with a decorator on
the pseudocode function.

One practical ramification of this change is that the `instructions`
dict no longer contains mappints of instruction names to (fn,encoding)
pairs. Instead, the mapping is from name to function and each function
is annotated with its encoding (which is an object).