Add descriptive comments on pseudocode functions.
In order to generate more informative descriptions of these instructions
(e.g., LaTeX pages), incorporate a description of each function in a
pre-function comment.

We take this approach rather than doc comments because doc comments are
interpreted as part of the function. When we want to print the
pseudocode of a function, we don't want the doc comment to be included.